created components folder moved card component into components folder
